+++ Runtime/Scripts/GameAnalytics.cs
@@ -245,6 +245,12 @@ namespace GameAnalyticsSDK
             {
                 SetEnabledManualSessionHandling(true);
             }
+
+            EnableSDKInitEvent(SettingsGA.EnableSDKInitEvent);
+            EnableFpsHistogram(SettingsGA.EnableFPSHistogram);
+            EnableMemoryHistogram(SettingsGA.EnableMemoryHistogram);
+            EnableHealthHardwareInfo(SettingsGA.EnableHardwareTracking);
+
         }
 
         public static void Initialize ()
@@ -1054,6 +1060,51 @@ namespace GameAnalyticsSDK
         public static long StopTimer(string key)
         {
             return GA_Wrapper.StopTimer(key);
+        }
+
+        // ----------------------- HEALTH EVENT --------------------------------------------//
+
+        /// <summary>
+        /// enable the SDK init event to automatically track the boot time (time from application launch to the GameAnalytics SDK initialization).
+        /// (Android & iOS ONLY)
+        /// </summary>
+        /// <param name="flag">should be enabled or not</param>
+        public static void EnableSDKInitEvent(bool flag)
+        {
+            GA_Setup.EnableSDKInitEvent(flag);
+        }
+
+        /// <summary>
+        ///  Enable FPS sampling across the entire session to ultimately send an FPS histogram via the Session Performance Event.
+        ///  (Android & iOS ONLY)
+        /// </summary>
+        /// <param name="flag">should be enabled or not</param>
+        public static void EnableFpsHistogram(bool flag)
+        {
+            GA_Setup.EnableFpsHistogram(flag);
+        }
+
+        /// <summary>
+        ///  Enable memory usage sampling across the entire session to ultimately send a memory usage histogram via the Session Performance Event.
+        /// (Android & iOS ONLY)
+        /// </summary>
+        /// <param name="flag">should be enabled or not</param>
+        public static void EnableMemoryHistogram(bool flag)
+        {
+            GA_Setup.EnableMemoryHistogram(flag);
+        }
+
+        /// <summary>
+        /// (EXPERIMENTAL) enable discovery of device hardware information like,
+        /// cpu model, number of cpu cores, GPU model, chipset/hardware (if available).
+        /// these data points are added as properties to existing health
+        /// events (error, SDK init, session performance) if those are enabled.
+        //  (Android & iOS ONLY)
+        /// </summary>
+        /// <param name="flag">should be enabled or not</param>
+        public static void EnableHealthHardwareInfo(bool flag)
+        {
+            GA_Setup.EnableHealthHardwareInfo(flag);
         }
 
         // ----------------------- IOS 14+ APP TRACKING TRANSPARENCY ---------------------- //